fix(containers): Fix containers pagination  (#32558)
### Pagination

 **Problem**
p-table expects ALL data to be loaded at once in the [value] property,
It then performs client-side pagination, showing only the first 40 items
per page

 **Fix**

Added `lazy=true` attribute to fix the pagination issue enabling
server-side pagination, now this is handled with the `onPage` property
which determines how to load the data, and the table uses [totalRecords]
to determine how many pages to show.

### Redundant API call

Removed `this.paginatorService.get()` as it was making an extra API call
while the p-table was already making one when loading the data.