feat(nextjs): Add YouTube component and adjust BannerCarousel (#32290)
This pull request adds a new `YouTube` component for embedding YouTube
videos and updates the styling of the `BannerCarousel` component to
improve image rendering. Additionally, the `YouTube` component is
integrated into the `pageComponents` object for usage across the
application.

### New Component Addition:

*
[`examples/nextjs/src/components/content-types/YouTube.js`](diffhunk://#diff-db2f0fb3d03f90b6f13fd5fd0ff90c299ca5477a1cdd9e94c3c2c9c88f399973R1-R33):
Introduced a new `YouTube` component to embed YouTube videos with
metadata such as title, author, length, and thumbnail. The component
includes responsive styling and accessibility features.

### Integration of the New Component:

*
[`examples/nextjs/src/components/content-types/index.js`](diffhunk://#diff-5912745dd6534531e269707b918724caf3562c2a7e17b935d46f78d9d72ea638R16):
Added the `YouTube` component to the list of imports and integrated it
into the `pageComponents` object for usage across the application.
[[1]](diffhunk://#diff-5912745dd6534531e269707b918724caf3562c2a7e17b935d46f78d9d72ea638R16)
[[2]](diffhunk://#diff-5912745dd6534531e269707b918724caf3562c2a7e17b935d46f78d9d72ea638R34)

### Styling Update:

*
[`examples/nextjs/src/components/content-types/BannerCarousel.js`](diffhunk://#diff-ebd2b2692ec4e32385eb48459386e0839b42b051faad9609463f4687bbed1c12L43-R43):
Updated the image styling in the `BannerCarousel` component to ensure
images are fully covered and rendered properly using `object-cover`.

Co-authored-by: Jalinson Diaz <zjaaaldev@gmail.com>