feat: Add API endpoint to check expiring tokens and notify (#32647)
- Created API endpoint to get tokens that are about to expire within a
configurable number of days.

This endpoint, for admin users, returns all expiring tokens from all
users.
For limited users, returns only their own expiring tokens.

The days are configured by a new config property
**`EXPIRING_TOKEN_LOOKAHEAD_DAYS`** (default 7).

The user requesting is authenticated as indicated in requirements.

- Also added a new functionality to notify users when API Tokens are
about to expire, using the same logic.

This functionality throws alerts when logging in, if the user is Admin
it shows "Some API Tokens are about to expire. Please review them here"
and a link to the API endpoint
If the user is a Limited User "You have API Tokens that are about to
expire. Please let your Administrator know about this."

The alerts can be disabled from a new config property
**`DISPLAY_EXPIRING_TOKEN_ALERTS`** (default true)


### **Alert on admin user**


https://github.com/user-attachments/assets/0c875827-f8ce-4b18-b467-357395f04fe5


### **Alert on limited user**



https://github.com/user-attachments/assets/21a0c0ff-d36c-4dd7-b679-e4916eb983dd