Fixed workflow dropdown on content search (#32900)
https://github.com/user-attachments/assets/e6717e5f-784f-418b-9a37-120ee213d00a

This issue was about a misstime between the load from the API and the
load from the component.
Basically, sometimes the API load before the component exists, so the
component never was updated.
And sometimes the API load after, thats why the isse was intermitent.

With this change, the component always is loaded before the hydrate the
options inside

This PR fixes: #32899

Co-authored-by: Kevin <kfariid@gmail.com>