feat(content edit) #31734 : Create a REST Endpoint for returning the History data (#32906)
### Proposed Changes
* This PR introduces a new REST Endpoint to the
`ContentVersionResource`, called `/id/{identifier}/history`. It allows
you to retrieve a paginated minified set of versions that belong to a
specific Contentlet based on its Identifier. This was done by adding a
new Transform Option and its respective implementation, following our
existing code pattern.
* The retrieved information can be used for building the `History` tab
in the Content Editor window.
* A new implementation using the dedicated
`ResponseEntityPaginatedDataView` was added to the `PaginationUtil`
class. This allows us to move away from returning the generic `Response`
object, and use the Swagger annotations as expected.
* New method overloads for the `ContentletAPI` and the
`ContentletFactory` classes allow you to paginate the Contentlet version
data. Considering that the number of versions of a given Contentlet can
be extremely high, having this feature in place provides developers more
control on how they can retrieve data more safely and efficiently.
* A new Postman Collection was added for the `ContentVersionResource`,
as we didn't have any tests for it.

This PR fixes: #31734