fix(Content Import) Stop On Error inconsistent Behavior Refs:#32416 (#32767)
The goal of this PR was to ensure that errors in our import process
produce a meaningful, user-friendly error message along with a
corresponding error code. Additionally, when the **stopOnError** flag is
set to true, the import process should halt on the first error.

Previously, the **stopOnError** flag was being incorrectly defaulted to
true; this behavior has been fixed so that the default is now false, as
intended.

**Improvements and Changes**
• A comprehensive test has been added that covers various field types.
It intentionally triggers errors and verifies that the responses include
coherent error messages and the correct error codes.
• Some existing error messages were improved, and exceptions now include
more contextual information that can help users understand the issue.
• A new ImportLineError interface was introduced, allowing additional
metadata to be attached to exceptions raised during import.
• The class FieldHandlerStrategyFactory, which is responsible for
assigning values to fields based on their type, has been modified to
include additional validation logic and improved error handling.
• A fix that validated if an expression was a Lucene Query was made. As
the old code had some deficiencies and would interpret anything with a
hypen a valid lucene query

**Notes for Testers**
• Trying to import a Category that doesn’t exist will result in an
error.
• In contrast, importing a Tag that doesn’t exist will result in that
tag being created automatically.
• Many fields currently lack validation and will accept nearly any
value. For example:
	•	Text and Custom fields accept arbitrary input.
• A Radio field expects a boolean value, and the following values are
interpreted as true: 1, true, t, yes, on. Any other value will be
interpreted as false.
• A Multi-select list field will accept any chunk of text as a valid
value. However, the expected format is a comma-separated list of valid
entries.

**Additional Notes**

Although a StoryBlock field is stored as a JSON object internally, the
system currently accepts any input during import. This is intentional—it
allows clients to use endpoints to create content without being forced
to submit fully structured values. When edited through the UI,
StoryBlock fields are saved in the correct structured format.