chore: implement 404 strategy and remove logs (#32998) (#33001)
As discussed on the parent
[issue](https://github.com/dotCMS/core/issues/32430#issuecomment-3185510872)
a 404 strategy was added, and now the addPermissionsToCache function
doesn't log an error if the list of permissions is empty, it returns an
empty collection and puts on the cache a new object
`EmptyPermissionsCacheEntry` which will represent this 404 and expires
in 5 minutes.

Then when getting the permissions from the cache if it is this object
and it has expired then remove it from the cache and return null to
recalculate it.

This PR fixes: #32998

---------

Co-authored-by: Daniel Silva <daniel.silva@dotcms.com>