fixes(unique fields) #32765 : [Unique Fields] Upgrade with inconsistencies leave the system unusable (#33018)
### Proposed Changes
* Includes a new step in the Database Unique Field Validation process
where duplicate records are fixed and reported instead of having dotCMS
start up with an invalid `unique_fields` table.
* Records with the same hash belong to Contentlets that have the same
unique value with different casing. This is NOT correct. So, what the
code does now is:
* Remove the primary key constraint from the DB table, and load the
unique value data as usual.
  * Leave one "as is" and add a "prefix" to the other ones.
  * Re-generate the hash in order to have a different value.
  * Bring the primary key back and let dotCMS start as usual.
* Generate a log file with absolutely all records with conflicting data.
This way, users can inspect it and manually fix the corrupted
Contentlets.
* Such a log file can be displayed and downloaded from the `Logs`
portlet in the back-end.

This PR fixes: #32765