fixes(unique fields) #32309 : [Unique Fields] : Same unique value inside a Variant must be supported (#33045)
### Proposed Changes
* Makes sure that unique field values inside a Variant are supported.
* When a Variant is saved/published, its unique value will be checked
against its respective default version. If the value is the same, the
validation is skipped and no record will be added to the unique_fields
table.
* When a Variant is saved/published and its value is different from the
one in the default version, the validation process will be executed as
usual.
* If you update the unique value in a default version, and have already
created Variants of it for an Experiment, they must go ahead and
manually update such Variants as well.
* Variants of a default version should NOT match unique values for other
Contentlets outside the Experiment. If they do, once they get promoted,
an error will be thrown as it won’t be possible to publish such
Contentlets, and they must be manually fixed.