fix: Fix Reset Approvals sub action (#33112)
**Problem**

The JSON created as a description to detail the action and add it to the
history wasn't building in the correct format as it was using single
quotes`'` instead of double quotes `"`. So when approving and trying to
get the data from the history after a reset of the approvers, the parse
of the data wasn't accurate and it wasn't getting the `RESET` state as
it should.

The fix was to change the single quotes to double. Now the Reset
Approvals sub action works as it should.

Also, a test was added to improve coverage

This PR fixes: #33031