fix: fix related content field removal (#33230) (#33263)
The field was being set to null to prevent an infinite recursion issue.
This problem only occurs with contents that are related to another
content of the same type through a RelationshipField or StoryBlockField.
If the child (or related) content has a field ( of a type that is not at
risk of generate a recursion issue) with the same name, we now avoid
setting that field to null. This fixes the issue while still preserving
the functionality that prevents infinite recursion.

This PR fixes: #33230