feat: landings in epic template
We noticed in 2025 that tracking landings has been hit or miss.  This proposal includes a new field in the epic issue that is lighter weight than the reverse press release, but is intended to spark the conversation needed to have a clear understanding of what should happen if we've built the feature correclty.