Issue 33114 UUID mismatch on page render (#33357)
### **Problem**

The render of the page was returning the uuid of the container as
`LEGACY_RELATION_TYPE`, this was generating problems on the UVE and the
way it handles the position of the containers.

### **Fix**

The fix was to transform on the response the values of the uuids that
were `LEGACY_RELATION_TYPE` to a default one, this way the new
transformed uuids will match with the ones on the layout and the UVE
will work as it should.