[30939] feat: improve swagger documentation of job queues resources (#33300)
This pull request introduces a new schema and documentation for handling
multipart job queue parameters in the API. The most important changes
include the addition of a JobParamsSchema class to define the structure
of the job parameters and a JobQueueDocs class to provide documentation
for the API endpoint.

Additions to API schema and documentation:

[dotCMS/src/main/java/com/dotcms/rest/api/v1/job/JobParamsSchema.java](https://github.com/dotCMS/core/pull/32504/files#diff-948cf1a3843b79cf0b8bf6de2f8d86cd96b9e6c3d8905f7649fe061c2568d98fR1-R46):
Added a new class, JobParamsSchema, to define the schema for multipart
job queue parameters. This includes fields for a file (file) and
job-specific parameters (params), along with their respective getters
and setters.


[dotCMS/src/main/java/com/dotcms/rest/api/v1/job/JobQueueDocs.java](https://github.com/dotCMS/core/pull/32504/files#diff-8c562471b62a91702ff765f40a3c4e9779518a9761dfd019f2b17b123f7cd240R1-R19):
Added a new class, JobQueueDocs, to provide documentation for the API
endpoint. This includes details about the multipart/form-data fields
(file and form) and an example JSON for the form field.

<img width="566" height="836" alt="image"
src="https://github.com/user-attachments/assets/ac86bcf2-4c16-42f1-8c0a-ec9ee4739621"
/>

This PR fixes: #30939