feat(permission): allow category field  as secondary permission check for contentlets (#33389)
Introduced a method to enforce optional category-level permissions at
the contentlet level, ensuring users have access to associated
categories before accessing the content.

This introduces a method into the PermissionBitApi that checks if the
content that is being permissioned has any category fields with the
field variable `secondaryPermissionCheck=true`. If this field variable
is set on the content type, then the PermissionsAPI will also check to
insure that the user accessing the content has access to one or more of
the categories that have been used to categorize the content.

Rule for access:
- If the category field is empty or user has access to at least 1 of the
categories set in the field, then they can access the content.
- If the user does not have access to any of the categories that have
been used to categorize the content, then they will not be able to
access the content.
- if multiple category fields are found with the field variable
`secondaryPermissionCheck=true` then each one will be checked based on
the rules above. If any of the category fields fail access, then the
user will not be able to access the content.

In truth, the rules above are more complex to write than to use. Set a
field as `secondaryPermissionCheck=true` and then only users with access
to specific categories will be able to view and access the categorized
content.

### Entire Behavior can be turned off by setting:

`DOT_PERMISSION_SECONDARY_CATEGORY_CHECK=false`







ref: #33388