fix: add params to the request attribute (#33479)
For the Oauth scenario described in the ticket, the problem is that the
plugin won't get any queryString value because the url has the referrer
value and the params are removed from it, so now before the Oauth
interceptor executes, the queryString value is being set on a session
attribute, this way the plugin is going to be able to get this value and
build the url correctly.

To achieve this and verify the fix, we should update the Oauth plugin on
the `sendAuthorization` function on `LoginRequiredOAuthInterceptor.java`
for something like this

```
private void sendForAuthorization(final HttpServletRequest request,
            final HttpServletResponse response,
            final OAuthService service,
            final DefaultApi20 apiProvider) throws IOException {

        String retUrl = (String) request.getAttribute(JAVAX_SERVLET_FORWARD_REQUEST_URI);

        if (request.getSession().getAttribute(OAUTH_REDIRECT) != null) {
            retUrl = (String) request.getSession().getAttribute(OAUTH_REDIRECT);
        }

        if (request.getParameter(REFERRER) != null) {
            retUrl = request.getParameter(REFERRER);
        }

        String queryString = (String) request.getSession().getAttribute("javax.servlet.forward.query_string");

        StringBuilder urlBuilder = new StringBuilder(retUrl);
        if (queryString != null && !queryString.isEmpty()){
                urlBuilder.append("?").append(queryString);
        }
        retUrl = urlBuilder.toString();

        request.getSession().setAttribute(OAUTH_REDIRECT, retUrl);

     ....... code continues here
```


For the SAML scenario, the problem may be the `REDIRECT_AFTER_LOGIN`
attribute being lost on the session, maybe because of a session lost or
it never was set correctly in it. And then on the SAML interceptor,
checking that the relayState property is not present and trying to
fallback to this property it will find it null and the redirect will end
on `/dotAdmin/` when it shouldn't. So this property was also added to
the session attribute in the previous step before the Saml interceptor
executes. This **should** allow the interceptor to get the correct
value, but as we can't test it we are not able to verify this.
Apart from that, a small improvement was implemented to keep the
queryString params value on the redirect url.

This PR fixes: #33245