feat(Content Drive): Add ensure param when retrieving Content Types in the filter (#33576)
# Summary:

Ensures that when we share/reload a page with selected **Content Types**
in the filter, we keep them even though the Content Types selected are
not in the initial default pagination when the page is loaded for the
first time.


https://github.com/user-attachments/assets/aba885c6-1cb4-4d80-af49-55c0f6598fd0

> ℹ️ **Note:** I made the filter to bring just 10 Content Types but it
is just for demo purposes, the normal behaviour is to bring 40 Content
Types.

This PR fixes: #33152