Issue 31827 content import fails using site name (#33648)
### **Problem**

The import was failing because the `Site` field sent in the payload
contained the Site **name** instead of the Site **identifier**.
When looking up the existing contentlet by key, this mismatch caused the
import to fail and resulted in a new, incorrect contentlet being
created.

### **Fix**

We now convert the Site name to its corresponding Site identifier to
ensure we are validating with the correct values. This ensures that when
we try to import using the `Site or Folder` field as key we can use
either the identifier or the name and the import will work.

This PR fixes: #31827