fix(view-browser): Remove unnecessary long word shortening for asset name (#33713)
### Proposed Changes
This pull request makes a minor adjustment to the way asset names are
displayed in the browser view. The change removes the use of the
`shortenLongWords` function on the asset name, which previously
truncated long words in the name to prevent them from breaking the
column layout.

* Removed the call to `shortenLongWords` on the `name` variable, so
asset names will no longer have individual long words truncated in the
browser view.

### Checklist
- [ ] Tests
- [ ] Translations
- [ ] Security Implications Contemplated (add notes if applicable)




This PR fixes: #32802