fix(cache): fixing a condition where the old threadexecution pool would be killed on a cache flush and never come back. (#33805)
This code changes the H22Cache to reinit the threadpool if it has been
shutdown (which happens when you do a full cache flush in dotCMS).

Otherwise, when you flush cache, the pool is killed dead and the cache
still tries to use it for async operations.

Oh, it also adds a useful filter when dumping threads that will ignore
any threads that are not dotCMS based threads, e.g. reactor and tomcat
threads.



ref: #33784