add styleProperties field into MultiTree model and its implementations (#33731)
## Sumary

Implement database persistence for style properties by modifying the
database schema and updating the MultiTree data model to support the new
styling functionality.

### Proposed Changes
- [x] Add `style_properties` field into `multi_tree` table using the
`postgres.sql` for new databases.
- [x] Add `style_properties` field into `multi_tree` table using an
Update Task for existing databases.
- [x] Integration tests for the Update Task.
- [x] Add new integration test into the Suite a3.

Plus add the connection between the database and Java clases to manage
the new `styleProperties` field.

### Proposed Changes
- [x] `MultiTree` Add styleProperties class field (constructor, getter
and setter).
- [x] `MultiTreeAPIImpl` Add new field into the insert query and
operations validating that the value is a valid Json String.
- [x] `MultiTreeAPITest` Test different scenarios where the
styleProperties is present, updated or not defined.
- [x] `MultiTreeTransformer` Includes MultiTree field.


This PR fixes: #33693