feat(edit mode) #33988 : Check if Custom Fields have access to specif… (#34004)
### Proposed Changes
* Exposes a new method in the existing `structures` ViewTool that allow
developers to check whether the Content Type that the ViewTool is called
in has the new Edit mode enabled or not.
* This mechanism is going to be used by the Angular layer to determine
how existing Custom Fields are going to be rendered in the new Content
Edit Mode. That is, the mode that uses Angular instead of the legacy
JSP/Dojo UI.

This PR fixes: #33988